Move some functions to separate modules
- Move console output behavior to reports.js
- Move utility functions to utils.js
- Make fse a real dependency
