+++ packages/truffle-workflow-compile/legacy/index.js
@@ -1,52 +1,21 @@
 const debug = require("debug")("workflow-compile");
 const mkdirp = require("mkdirp");
 const { promisify } = require("util");
-const Config = require("truffle-config");
 const solcCompile = require("truffle-compile/legacy");
 const vyperCompile = require("truffle-compile-vyper");
 const externalCompile = require("truffle-external-compile");
-const expect = require("truffle-expect");
-const Resolver = require("truffle-resolver");
-const Artifactor = require("truffle-artifactor");
-const OS = require("os");
+const { prepareConfig, multiPromisify } = require("../utils");
+const {
+  reportCompilationStarted,
+  reportNothingToCompile,
+  reportCompilationFinished
+} = require("../reports");
 
 const SUPPORTED_COMPILERS = {
   solc: solcCompile,
   vyper: vyperCompile,
   external: externalCompile
 };
-
-function prepareConfig(options) {
-  expect.options(options, ["contracts_build_directory"]);
-
-  expect.one(options, ["contracts_directory", "files"]);
-
-  // Use a config object to ensure we get the default sources.
-  const config = Config.default().merge(options);
-
-  config.compilersInfo = {};
-
-  if (!config.resolver) config.resolver = new Resolver(config);
-
-  if (!config.artifactor) {
-    config.artifactor = new Artifactor(config.contracts_build_directory);
-  }
-
-  return config;
-}
-
-function multiPromisify(func) {
-  return (...args) =>
-    new Promise((accept, reject) => {
-      const callback = (err, ...results) => {
-        if (err) reject(err);
-
-        accept(results);
-      };
-
-      func(...args, callback);
-    });
-}
 
 const Contracts = {
   collectCompilations: async compilations => {
@@ -94,7 +63,7 @@ const Contracts = {
 
       if (numberOfCompiledContracts === 0) this.reportNothingToCompile(options);
 
-      this.reportCompilationFinished(options, config);
+      this.reportCompilationFinished(config);
       const result = await this.collectCompilations(compilations);
       if (callbackPassed) return callback(null, result);
       return result;
@@ -133,44 +102,9 @@ const Contracts = {
     );
   },
 
-  reportCompilationStarted: options => {
-    const logger = options.logger || console;
-    if (!options.quiet) {
-      logger.log(OS.EOL + `Compiling your contracts...`);
-      logger.log(`===========================`);
-    }
-  },
-
-  reportCompilationFinished: (options, config) => {
-    const logger = options.logger || console;
-    const { compilersInfo } = config;
-    if (!options.quiet) {
-      if (Object.keys(compilersInfo).length > 0) {
-        logger.log(
-          `> Artifacts written to ${options.contracts_build_directory}`
-        );
-        logger.log(`> Compiled successfully using:`);
-
-        const maxLength = Object.keys(compilersInfo)
-          .map(name => name.length)
-          .reduce((max, length) => (length > max ? length : max), 0);
-
-        for (const name in compilersInfo) {
-          const padding = " ".repeat(maxLength - name.length);
-
-          logger.log(`   - ${name}:${padding} ${compilersInfo[name].version}`);
-        }
-      }
-      logger.log();
-    }
-  },
-
-  reportNothingToCompile: options => {
-    const logger = options.logger || console;
-    if (!options.quiet) {
-      logger.log(`> Everything is up to date, there is nothing to compile.`);
-    }
-  },
+  reportCompilationStarted,
+  reportCompilationFinished,
+  reportNothingToCompile,
 
   writeContracts: async (contracts, options) => {
     await promisify(mkdirp)(options.contracts_build_directory);
