+++ packages/truffle-workflow-compile/package.json
@@ -4,6 +4,7 @@
   "description": "Core workflow behavior for `truffle compile` command",
   "dependencies": {
     "mkdirp": "^0.5.1",
+    "fs-extra": "^7.0.1",
     "truffle-artifactor": "^4.0.25",
     "truffle-compile": "^4.1.6",
     "truffle-compile-vyper": "^1.0.23",
@@ -29,7 +30,6 @@
   "gitHead": "b207efb3c1409746537293b3e0fc27350029188e",
   "devDependencies": {
     "debug": "^4.1.1",
-    "fs-extra": "^7.0.1",
     "mocha": "5.2.0"
   }
 }
